STY: relocate remove_preamble for icon instruments
